requirements.txt
twilio 
python-dotenv
selenium 
requests
datetime
python-dotenv
google-generativeai
google 
google.cloud